--- Packages/KeyAppKit/Sources/BankTransfer/StrigaProvider/Models/UserDetailsResponse.swift
@@ -1,3 +1,0 @@
-import Foundation
-
-public typealias UserDetailsResponse = RegistrationData
