+++ Packages/KeyAppKit/Sources/BankTransfer/StrigaRemoteProvider/Models/File.swift
@@ -1,0 +1,6 @@
+import Foundation
+
+public enum StrigaProviderError: Error {
+    case invalidRequest(String)
+    case invalidResponse
+}
