+++ p2p_wallet/Scenes/Main/BankTransfer/StrigaRegistration/FirstStep/StrigaRegistrationFirstStepViewModel.swift
@@ -36,7 +36,7 @@ final class StrigaRegistrationFirstStepViewModel: BaseViewModel, ObservableObjec
     var fieldsStatuses = [Field: StrigaRegistrationTextFieldStatus]()
 
     @Published var selectedCountryOfBirth: Country?
-    @Published private var dateOfBirthModel: RegistrationData.DateOfBirth?
+    @Published private var dateOfBirthModel: UserDetailsResponse.DateOfBirth?
     private let dateFormat = "dd.mm.yyyy"
 
     init(country: Country) {
@@ -63,7 +63,7 @@ final class StrigaRegistrationFirstStepViewModel: BaseViewModel, ObservableObjec
             .map { $0.split(separator: ".") }
             .map { [weak self] components in
                 if components.count == self?.dateFormat.split(separator: ".").count {
-                    return RegistrationData.DateOfBirth(year: Int(components[2]), month: Int(components[1]), day: Int(components[0]))
+                    return UserDetailsResponse.DateOfBirth(year: Int(components[2]), month: Int(components[1]), day: Int(components[0]))
                 } else {
                     return nil
                 }
@@ -90,7 +90,11 @@ private extension StrigaRegistrationFirstStepViewModel {
     func fetchSavedData() {
         Task {
             do {
-                let data = try await service.getRegistrationData()
+                guard let data = try await service.getRegistrationData() as? UserDetailsResponse
+                else {
+                    throw StrigaProviderError.invalidResponse
+                }
+                
                 await MainActor.run {
                     email = data.email
                     phoneNumber = data.mobile.number
@@ -171,11 +175,11 @@ private extension StrigaRegistrationFirstStepViewModel {
             .sinkAsync { [weak self] contacts, credentials, dateOfBirth in
                 guard let self else { return }
 
-                try? await self.service.save(data: RegistrationData(
+                try? await self.service.save(data: UserDetailsResponse(
                     firstName: credentials.0,
                     lastName: credentials.1,
                     email: contacts.0,
-                    mobile: RegistrationData.Mobile(countryCode: "", number: contacts.1),
+                    mobile: UserDetailsResponse.Mobile(countryCode: "", number: contacts.1),
                     dateOfBirth: dateOfBirth.0,
                     placeOfBirth: dateOfBirth.1?.code
                 ))
